backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
sqlalchemy
python-dotenv
alembic
cryptography